Refactor table into separate component
